Refactor ModelInflator to database package
